LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-window-open-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,%3Chtml%3E%3Cbody%3E%3Cp%3EOpened%20Frame%3C/p%3E%3Cp%20id='accessMe'%3E%3C/p%3E%3Ciframe%3E%3C/iframe%3E%3C/body%3E%3C/html%3E from frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-to-data-url-window-open.html. Domains, protocols and ports must match.

Opener Frame

PASS: Access to a window opened with a data: URL was denied.
